feat(discovery): promote terminal/client/profile discovery to pkg/discovery
